etl_pipeline/requirements.txt
# ETL pipeline dependencies
# Install inside the custom Airflow image (see Dockerfile)

# PostgreSQL driver (source DB)
psycopg2-binary>=2.9.9

# Vector store client
chromadb>=0.5.0

# Embedding providers (LangChain ecosystem)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-openai>=0.1.0           # only needed when EMBEDDING_PROVIDER=openai

# HuggingFace local embeddings (default provider)
sentence-transformers>=3.0.0

# Utilities
python-dotenv>=1.0.0
